자신의 글만 삭제 가능

+++ src/main/java/com/example/prjjsp20241022/service/BoardService.java
@@ -68,9 +68,14 @@ public class BoardService {
         return mapper.selectById(id);
     }
 
-    public void remove(Integer id) {
+    public void remove(Integer id, Member member) {
+        Board board = mapper.selectById(id);
+        if (board.getWriter().equals(member.getId())) {
+            mapper.deleteById(id);
+        } else {
+            throw new RuntimeException("삭제 권한이 없습니다.");
+        }
 
-        mapper.deleteById(id);
     }
 
     public void update(Board board) {
